WIP py/objroproxy: Implement sys.roproxy class.
Allows to wrap any object to provide a read-only view of it, where
attribute and subscript store operation are disabled, while all other
operations are forwarded to the underlying object.

Conceptually, this is a generalization of CPython's types.MappingProxyType:
https://docs.python.org/3/library/types.html#types.MappingProxyType .
But the latter intended to proxy only mappings, while roproxy is
suitable for sequences, mappings, objects, modules, etc.

Change-Id: Idb287fae3792be5bf84475ab103ce55b5ae01680
TODO: Make configurable.